requirements.txt
loguru==0.7.3
mili-ui==1.0.1
pygame-ce==2.5.2
requests==2.32.3
